Declare workspaceFolders support in server capabilities
This is the recommended way to declare up-front that the server
supports workspace folders and is interested in change notifications.
Without this server capability set, the vscode-languageserver-node
dependency tries to register for workspace change notifications
dynamically with a 'client/registerCapability' request. If the client
doesn't support dynamic capability registration and responds to that
request with a MethodNotHandled JSONRPC error, the promise for the
request is rejected which causes the server to exit.