feat: enable auto import behaviour
Enable auto import by creating a real TS project and project service
proxing the project methods to the languageServerHost.

As we use a real project behind the scenes the language host no longer
needs the auto import cache package to decorate as the project itself
provides the module specifier and import caching behaviour.

fixes: #19
fixes: vuejs/language-tools/issues#963
fixes: vuejs/language-tools/issues#1738