fix(language-server): add typescript and locale initialization options
This allows people to load TypeScript as suggested in #140. Since
`initializationOptions` is of type `any` there, it doesn’t really
matter. But the `InitializationOptions` interface is used by the test
utils and prohibits the user from passing additional properties.

Also it’s nice if all Volar based language servers that do allow the
client to pass the TypeScript SDK, do so in a consistent manner.