fix(language-core): handle empty mappings in updateVirtualCodeMapOfMap
If virtual code has no mappings, we still need to update virtual code
maps based on a source of undefined. Without this, the TypeScript plugin
isn’t able to find virtual code without mappings. For completions this
means it falls back to TypeScript’s own `getCompletsionsAtPosition`
implementation, which returns a list of everything in scope.

Refs https://github.com/mdx-js/mdx-analyzer/issues/427